endtoend/pyproject.toml
[tool.poetry]
name = "endtoend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.2.2"}
pydantic = "*"
langchain = "*"
langchain-community = "*"
langchain-openai = "*"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
